chore: think i've finished the refactor of this bloody component now
