+++ plugins/scaffolder/src/components/fields/RepoUrlPicker/index.ts
@@ -13,5 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { RepoUrlPicker } from './NewRepoUrlPicker';
+export { RepoUrlPicker } from './RepoUrlPicker';
 export { repoPickerValidation } from './validation';
